Refactor and reimplement  data storage for bones
Updates bones to be processed the same way as lights, meshes, nodes, and animations are handled.

Why we are doing this:
We intend to fix all bones not being instanced, as leaf bones are important in game engines. Without a mesh the bone is removed which should not be the case and is not expected behaviour.